update fields api-depot types
