+++ apps/api/src/modules/base_locale/sub_modules/habilitation/habilitation.controller.ts
@@ -18,7 +18,7 @@ import {
   ApiBearerAuth,
 } from '@nestjs/swagger';
 
-import { Habilitation } from '@/shared/modules/api_depot/types/habilitation.type';
+import { Habilitation } from '@/shared/modules/api_depot/api-depot.types';
 
 import { CustomRequest } from '@/lib/types/request.type';
 import { AdminGuard } from '@/lib/guards/admin.guard';
