+++ apps/cron/src/tasks/detect_conflict.task.ts
@@ -8,7 +8,7 @@ import {
   StatusBaseLocalEnum,
 } from '@/shared/entities/base_locale.entity';
 import { ApiDepotService } from '@/shared/modules/api_depot/api_depot.service';
-import { Revision } from '@/shared/modules/api_depot/types/revision.type';
+import { Revision } from '@/shared/modules/api_depot/api-depot.types';
 
 import { Task } from '../task_queue.class';
 import { CacheService } from '@/shared/modules/cache/cache.service';
