--- libs/shared/src/modules/api_depot/types/file.type.ts
@@ -1,12 +1,0 @@
-export enum TypeFileEnum {
-  BAL = 'bal',
-}
-
-export type File = {
-  id?: string;
-  revisionId?: string;
-  size?: number;
-  hash?: string;
-  type?: TypeFileEnum;
-  createdAt?: Date;
-};
